+++ TabletopTweaks/Bugfixes/Classes/Demon.cs
@@ -2,8 +2,11 @@
 using Kingmaker.Blueprints;
 using Kingmaker.Blueprints.Classes;
 using Kingmaker.Blueprints.JsonSystem;
+using Kingmaker.Designers.Mechanics.Facts;
 using Kingmaker.UnitLogic.Buffs.Blueprints;
 using Kingmaker.UnitLogic.FactLogic;
+using Kingmaker.UnitLogic.Mechanics;
+using Kingmaker.UnitLogic.Mechanics.Properties;
 using TabletopTweaks.Config;
 using TabletopTweaks.Extensions;
 using TabletopTweaks.NewComponents;
@@ -21,6 +24,7 @@ namespace TabletopTweaks.Bugfixes.Classes {
                 Main.LogHeader("Patching Demon Resources");
 
                 PatchBalorVorpalStrike();
+                PatchBrimorakAspect();
 
                 void PatchBalorVorpalStrike() {
                     if (ModSettings.Fixes.Demon.IsDisabled("BalorVorpalStrike")) { return; }
@@ -39,8 +43,31 @@ namespace TabletopTweaks.Bugfixes.Classes {
                     });
                     BalorVorpalStrikeFeature.AddComponent<RecalculateOnEquipmentChange>();
 
-                    Main.LogPatch("Patched", BalorVorpalStrikeFeature);
-                    Main.LogPatch("Patched", BalorVorpalStrikeBuff);
+                    Main.LogPatch(BalorVorpalStrikeFeature);
+                    Main.LogPatch(BalorVorpalStrikeBuff);
+                }
+
+                void PatchBrimorakAspect() {
+                    if (ModSettings.Fixes.Demon.IsDisabled("BrimorakAspect")) { return; }
+
+                    var BrimorakAspectEffectBuff = Resources.GetBlueprint<BlueprintBuff>("f154542e0b97908479a578dd7bf6d3f7");
+                    var BrimorakAspectEffectProperty = Resources.GetBlueprintReference<BlueprintUnitPropertyReference>("d6a524d190f04a7ca3f920d2f96fa21b");
+
+                    BrimorakAspectEffectBuff.TemporaryContext(bp => { 
+                        bp.RemoveComponents<DraconicBloodlineArcana>();
+                        bp.AddComponent<BonusDamagePerDice>(c => {
+                            c.CheckDescriptor = false;
+                            c.SpellsOnly = true;
+                            c.UseContextBonus = true;
+                            c.Value = new ContextValue() {
+                                ValueType = ContextValueType.CasterCustomProperty,
+                                m_CustomProperty = BrimorakAspectEffectProperty,
+                                ValueRank = Kingmaker.Enums.AbilityRankType.StatBonus
+                            };
+                        });
+                    });
+
+                    Main.LogPatch(BrimorakAspectEffectBuff);
                 }
             }
         }